@@ -1,7 +1,7 @@
 class CommentModel{
-  String name;
-  String date;
-  String comment;
+  String? name;
+  String? date;
+  String? comment;
   CommentModel({
     this.name,
     this.date,
