@@ -1,5 +1,6 @@
-import 'package:cat_adoption/screens/homeScreen.dart';
+import 'package:cat_adoption/landingScreen.dart';
 import 'package:flutter/material.dart';
+import 'package:get/get.dart';
 
 void main() {
   runApp(MyApp());
@@ -9,11 +10,10 @@ class MyApp extends StatelessWidget {
   // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
+    return GetMaterialApp(
       title: 'Cat Adoption',
       theme: ThemeData(
         fontFamily: 'Circe',
-        // fontFamily: 'Vialog',
         primarySwatch: Colors.blue,
         visualDensity: VisualDensity.adaptivePlatformDensity,
         scaffoldBackgroundColor: Colors.white,
@@ -25,7 +25,7 @@ class MyApp extends StatelessWidget {
           )
         )
       ),
-      home: HomeScreen(),
+      home: LandingScreen(),
     );
   }
 }
